Fix PartialConfiguration by converting it to Scala (automatically) XXX please fix!
That's a workaround for a crazy javac-Scalac interaction (probably a bug
which people don't really care about). Explanation below.
The automatic conversion is done by IDEA and the result is ugly,
but works apparently.

Symptoms:

scalac <classpath> <path/to>/PartialConfiguration.java completes successfully.
scalac <classpath> <path/to>/PartialConfiguration.java
<path/to>/PartialConfigurationParser.scala fails with the same message,
complaining that AbstractFeatureExprModule.AbstractFeatureExpr is not a valid
type because AbstractFeatureExprModule is not a value. That would be a valid
error message if AbstractFeatureExprModule.AbstractFeatureExpr would appear in
a Scala program.

Explanation:

PartialConfiguration.java is needed by PartialConfigurationParser.scala, and
to support that the Scala compiler must also parse PartialConfiguration.java
(or at least that's the current config). Doing this parsing, however, scalac
does not understand AbstractFeatureExprModule.AbstractFeatureExpr (which is
needed at the interface between Scala and Java code) as javac does, but as
a Scala type - and that's of course invalid.